requirements.txt
Django~=3.0.10
djangorestframework~=3.12.2